Move tests to new file
